refactor: add paths config
