+++ gatsby/node/create-pages.js
@@ -2,7 +2,7 @@ const path = require("path")
 const pages = require("../../config/pages")
 const slashify = require("../../src/utils/slashify")
 const getToc = require("../../src/utils/get-toc.js")
-const { POST_TEMPLATE_PATH } = require("../../src/utils/constants")
+const { POST_TEMPLATE_PATH } = require("../../config/paths")
 
 const createPages = async ({ actions: { createPage }, graphql, reporter }) => {
   const {
