+++ gatsby/node/on-create-page.js
@@ -1,7 +1,7 @@
 const path = require("path")
 const pages = require("../../config/pages")
 const slashify = require("../../src/utils/slashify")
-const { IMAGES_PATH } = require("../../src/utils/constants")
+const { IMAGES_PATH } = require("../../config/paths")
 
 const onCreatePage = ({ page, actions: { createPage, deletePage } }) =>
   Object.values(pages)
